Get rid of vestal versions
Since the published gem is so out of date, we need to include it from
another branch on git on every host app Gemfile.

That makes it complicated to implement and we don't really use
vestal_versions for anyithing.

Perhaps later on I can come back and write versioning for the blog.